rename : toast, modal 구조화
